Function that executes a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -15,7 +15,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int index;
+	unsigned long int index;
 
 	if (array != 0 && size > 0  && action != 0)
 	{
